refactor(funding-platform): decompose oversized application views and services
Split the application list and detail views into thin presentational
components backed by container hooks, and extract shared query-param and
query-key helpers so the previously oversized files drop back under their
size budgets.

- ApplicationListWithAPI: move filter/sort/URL-sync/reviewer state into
  useApplicationListFilters; extract ApplicationListStatsBar and
  ApplicationListFilterBar
- ApplicationDetailView: move data fetching, handlers, and derived flags
  into useApplicationDetailView (container/presenter split)
- fundingPlatformService: extract fundingApplicationParams helper and a
  shared requestApplicationsExport, removing duplicated param building
- useFundingPlatform: extract the QUERY_KEYS factory into
  fundingPlatformQueryKeys and fold reviewer-scoped stats into it
