+++ components/FundingPlatform/ApplicationList/ApplicationListWithAPI.tsx
@@ -1,16 +1,10 @@
 "use client";
 
-import { ArrowDownTrayIcon, FunnelIcon } from "@heroicons/react/24/outline";
 import { useMutation } from "@tanstack/react-query";
-// eslint-disable-next-line import/no-extraneous-dependencies
-import debounce from "lodash.debounce";
-import { usePathname, useRouter, useSearchParams } from "next/navigation";
 import pluralize from "pluralize";
-import { type FC, useCallback, useEffect, useMemo, useState } from "react";
+import { type FC, useCallback, useMemo } from "react";
 import { toast } from "react-hot-toast";
 import InfiniteScroll from "react-infinite-scroll-component";
-import { useAccount } from "wagmi";
-import { Button } from "@/components/Utilities/Button";
 import {
   useApplicationExport,
   useFundingApplications,
@@ -23,11 +17,12 @@ import { useProgramReviewers } from "@/hooks/useProgramReviewers";
 import type { IApplicationFilters } from "@/services/fundingPlatformService";
 import { useProgramPrompts } from "@/src/features/prompt-management";
 import type { IFundingApplication } from "@/types/funding-platform";
-import formatCurrency from "@/utilities/formatCurrency";
 import { getAIColumnVisibility } from "../helper/getAIColumnVisibility";
 import ApplicationInbox from "./ApplicationInbox";
 import { ApplicationList } from "./ApplicationList";
-import { ReviewerFilterDropdown } from "./ReviewerFilterDropdown";
+import { ApplicationListFilterBar } from "./ApplicationListFilterBar";
+import { ApplicationListStatsBar } from "./ApplicationListStatsBar";
+import { useApplicationListFilters } from "./useApplicationListFilters";
 
 interface IApplicationListWithAPIProps {
   programId: string; // Normalized program id (chainId suffix stripped)
@@ -49,69 +44,28 @@ const ApplicationListWithAPI: FC<IApplicationListWithAPIProps> = ({
   onApplicationHover,
   showStatusActions = false,
   initialFilters = {},
-  onStatusChange: parentOnStatusChange,
   isAdmin = false,
 }) => {
-  const router = useRouter();
-  const pathname = usePathname();
-  const searchParams = useSearchParams();
-  const { address } = useAccount();
-  const [myReviewsOnly, setMyReviewsOnly] = useState(true);
-  // Reviewer addresses selected in the "All Applications" dropdown (lowercased)
-  const [selectedReviewerAddresses, setSelectedReviewerAddresses] = useState<string[]>([]);
-
-  // Initialize filters from URL params (excluding page for infinite scroll)
-  const [filters, setFilters] = useState<IApplicationFilters>(() => {
-    const urlFilters = { ...initialFilters };
-    if (searchParams.get("search")) urlFilters.search = searchParams.get("search")!;
-    if (searchParams.get("status")) urlFilters.status = searchParams.get("status")!;
-    if (searchParams.get("dateFrom")) urlFilters.dateFrom = searchParams.get("dateFrom")!;
-    if (searchParams.get("dateTo")) urlFilters.dateTo = searchParams.get("dateTo")!;
-    // Remove page handling for infinite scroll
-    return urlFilters;
-  });
-
-  // Local state for search input (immediate UI updates)
-  const [searchInput, setSearchInput] = useState(filters.search || "");
-
-  const [sortBy, setSortBy] = useState<IApplicationFilters["sortBy"]>(() => {
-    const urlSortBy = searchParams.get("sortBy");
-    return (urlSortBy as IApplicationFilters["sortBy"]) || "status";
-  });
-
-  const [sortOrder, setSortOrder] = useState<IApplicationFilters["sortOrder"]>(() => {
-    const urlSortOrder = searchParams.get("sortOrder");
-    return (urlSortOrder as IApplicationFilters["sortOrder"]) || "asc";
-  });
-
-  // Debounced search function (waits 500ms after user stops typing)
-  const debouncedSearch = useMemo(
-    () =>
-      debounce((searchValue: string) => {
-        setFilters((prev) => ({ ...prev, search: searchValue || undefined }));
-      }, 500),
-    []
-  );
-
-  // Active reviewer filter, shared across the list, stats, and export:
-  // "My Applications" scopes to the current user; "All Applications" scopes to
-  // the reviewer(s) selected in the dropdown (if any).
-  const reviewerFilter = useMemo(
-    () => ({
-      reviewerAddress: myReviewsOnly && address ? address.toLowerCase() : undefined,
-      reviewerAddresses:
-        !myReviewsOnly && selectedReviewerAddresses.length > 0
-          ? selectedReviewerAddresses
-          : undefined,
-    }),
-    [myReviewsOnly, address, selectedReviewerAddresses]
-  );
+  const {
+    filters,
+    searchInput,
+    sortBy,
+    sortOrder,
+    myReviewsOnly,
+    setMyReviewsOnly,
+    selectedReviewerAddresses,
+    setSelectedReviewerAddresses,
+    queryParams,
+    hasActiveFilters,
+    handleFilterChange,
+    handleSearchChange,
+    handleSortChange,
+    clearFilters,
+  } = useApplicationListFilters(initialFilters);
 
   const {
     applications,
     total,
-    page,
-    totalPages,
     stats,
     isLoading,
     isFetchingNextPage,
@@ -119,14 +73,8 @@ const ApplicationListWithAPI: FC<IApplicationListWithAPIProps> = ({
     fetchNextPage,
     error,
     updateApplicationStatus,
-    isUpdatingStatus,
     refetch,
-  } = useFundingApplications(programId, {
-    ...filters,
-    sortBy,
-    sortOrder,
-    ...reviewerFilter,
-  });
+  } = useFundingApplications(programId, queryParams);
 
   // Load more function for infinite scroll
   const loadMore = useCallback(() => {
@@ -191,57 +139,6 @@ const ApplicationListWithAPI: FC<IApplicationListWithAPIProps> = ({
     }
   );
 
-  // Sync filters and sorting with URL
-  useEffect(() => {
-    const params = new URLSearchParams(searchParams.toString());
-
-    // Update filter params
-    if (filters.search) {
-      params.set("search", filters.search);
-    } else {
-      params.delete("search");
-    }
-
-    if (filters.status) {
-      params.set("status", filters.status);
-    } else {
-      params.delete("status");
-    }
-
-    if (filters.dateFrom) {
-      params.set("dateFrom", filters.dateFrom);
-    } else {
-      params.delete("dateFrom");
-    }
-
-    if (filters.dateTo) {
-      params.set("dateTo", filters.dateTo);
-    } else {
-      params.delete("dateTo");
-    }
-
-    // Remove page handling for infinite scroll
-
-    // Add sorting params
-    if (sortBy && sortBy !== "createdAt") {
-      params.set("sortBy", sortBy);
-    } else {
-      params.delete("sortBy");
-    }
-
-    // Always persist sortOrder in URL if it's set
-    if (sortOrder) {
-      params.set("sortOrder", sortOrder);
-    } else {
-      params.delete("sortOrder");
-    }
-
-    const queryString = params.toString();
-    const newUrl = queryString ? `${pathname}?${queryString}` : pathname;
-
-    router.replace(newUrl, { scroll: false });
-  }, [filters, sortBy, sortOrder, pathname, router, searchParams]);
-
   const statusChangeMutation = useMutation({
     mutationFn: (vars: {
       applicationId: string;
@@ -279,43 +176,9 @@ const ApplicationListWithAPI: FC<IApplicationListWithAPIProps> = ({
 
   const handleExport = useCallback(
     (format: "json" | "csv" = "json") => {
-      exportApplications(format, { ...filters, sortBy, sortOrder, ...reviewerFilter });
+      exportApplications(format, queryParams);
     },
-    [exportApplications, filters, sortBy, sortOrder, reviewerFilter]
-  );
-
-  const handleFilterChange = useCallback((newFilters: IApplicationFilters) => {
-    setFilters((prev) => ({ ...prev, ...newFilters }));
-  }, []);
-
-  const handleSearchChange = useCallback(
-    (value: string) => {
-      setSearchInput(value); // Update UI immediately
-      debouncedSearch(value); // Update filters after 500ms
-    },
-    [debouncedSearch]
-  );
-
-  // Cleanup debounced function on unmount
-  useEffect(() => {
-    return () => {
-      debouncedSearch.cancel();
-    };
-  }, [debouncedSearch]);
-
-  const handleSortChange = useCallback(
-    (newSortBy: string) => {
-      const typedSortBy = newSortBy as IApplicationFilters["sortBy"];
-      if (sortBy === typedSortBy) {
-        // Toggle sort order if clicking the same column
-        setSortOrder(sortOrder === "asc" ? "desc" : "asc");
-      } else {
-        // Set new sort column with default desc order
-        setSortBy(typedSortBy);
-        setSortOrder("desc");
-      }
-    },
-    [sortBy, sortOrder]
+    [exportApplications, queryParams]
   );
 
   // Show error state
@@ -341,198 +204,26 @@ const ApplicationListWithAPI: FC<IApplicationListWithAPIProps> = ({
     );
   }
 
-  const statsMap = [
-    {
-      title: "Total Applications",
-      value: stats?.totalApplications || 0,
-    },
-    {
-      title: "Pending Review",
-      value: stats?.pendingApplications || 0,
-    },
-    {
-      title: "Revision Requested",
-      value: stats?.revisionRequestedApplications || 0,
-    },
-    {
-      title: "Under Review",
-      value: stats?.underReviewApplications || 0,
-    },
-    {
-      title: "Approved",
-      value: stats?.approvedApplications || 0,
-    },
-    {
-      title: "Rejected",
-      value: stats?.rejectedApplications || 0,
-    },
-  ];
-
   return (
     <div className="w-full space-y-6">
-      {/* Statistics Bar */}
-      {stats && (
-        <div className="grid grid-cols-3 md:grid-cols-6 gap-4">
-          {statsMap.map((item) => (
-            <div
-              key={item.title}
-              className="bg-white dark:bg-zinc-800 p-4 rounded-lg border items-center justify-center"
-            >
-              <div className="text-2xl font-bold text-gray-900 dark:text-white text-center">
-                {formatCurrency(item.value || 0)}
-              </div>
-              <div className="text-sm text-gray-600 dark:text-gray-400 text-center">
-                {item.title}
-              </div>
-            </div>
-          ))}
-        </div>
-      )}
+      <ApplicationListStatsBar stats={stats} />
 
-      {/* Enhanced Filters */}
-      <div className="bg-white dark:bg-zinc-800 p-4 rounded-lg border">
-        <div className="grid grid-cols-1 md:grid-cols-4 gap-4">
-          <div>
-            <label
-              htmlFor="app-list-search"
-              className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2"
-            >
-              Search
-            </label>
-            <input
-              id="app-list-search"
-              type="text"
-              placeholder="Search by email, reference, or project title..."
-              value={searchInput}
-              onChange={(e) => handleSearchChange(e.target.value)}
-              className="w-full rounded-lg border border-gray-200 bg-white px-3 py-2 text-gray-900 placeholder:text-gray-300 dark:bg-zinc-800 dark:border-zinc-700 dark:text-zinc-100 dark:placeholder-zinc-300"
-            />
-          </div>
-
-          <div>
-            <label
-              htmlFor="app-list-status"
-              className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2"
-            >
-              Status
-            </label>
-            <select
-              id="app-list-status"
-              value={filters.status || ""}
-              onChange={(e) => handleFilterChange({ status: e.target.value })}
-              className="w-full rounded-lg border border-gray-200 bg-white px-3 py-2 text-gray-900 dark:bg-zinc-800 dark:border-zinc-700 dark:text-zinc-100"
-            >
-              <option value="">All Statuses</option>
-              <option value="pending">Pending</option>
-              <option value="resubmitted">Resubmitted</option>
-              <option value="under_review">Under Review</option>
-              <option value="revision_requested">Revision Requested</option>
-              <option value="approved">Approved</option>
-              <option value="rejected">Rejected</option>
-            </select>
-          </div>
-
-          <div>
-            <label
-              htmlFor="app-list-date-from"
-              className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2"
-            >
-              From Date
-            </label>
-            <input
-              id="app-list-date-from"
-              type="date"
-              value={filters.dateFrom || ""}
-              onChange={(e) => handleFilterChange({ dateFrom: e.target.value })}
-              className="w-full rounded-lg border border-gray-200 bg-white px-3 py-2 text-gray-900 dark:bg-zinc-800 dark:border-zinc-700 dark:text-zinc-100"
-            />
-          </div>
-
-          <div>
-            <label
-              htmlFor="app-list-date-to"
-              className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2"
-            >
-              To Date
-            </label>
-            <input
-              id="app-list-date-to"
-              type="date"
-              value={filters.dateTo || ""}
-              onChange={(e) => handleFilterChange({ dateTo: e.target.value })}
-              className="w-full rounded-lg border border-gray-200 bg-white px-3 py-2 text-gray-900 dark:bg-zinc-800 dark:border-zinc-700 dark:text-zinc-100"
-            />
-          </div>
-        </div>
-
-        <div className="flex justify-between mt-4 space-x-2">
-          <div className="flex flex-row items-center gap-2">
-            <div className="flex flex-row gap-1 items-center rounded-lg border border-gray-200 dark:border-zinc-700 p-1">
-              <button
-                type="button"
-                onClick={() => setMyReviewsOnly(true)}
-                aria-pressed={myReviewsOnly}
-                className={`px-3 py-1 text-sm font-medium rounded-md transition-colors ${
-                  myReviewsOnly
-                    ? "bg-blue-600 text-white"
-                    : "text-gray-600 dark:text-gray-400 hover:text-gray-900 dark:hover:text-white"
-                }`}
-              >
-                My Applications
-              </button>
-              <button
-                type="button"
-                onClick={() => setMyReviewsOnly(false)}
-                aria-pressed={!myReviewsOnly}
-                className={`px-3 py-1 text-sm font-medium rounded-md transition-colors ${
-                  !myReviewsOnly
-                    ? "bg-blue-600 text-white"
-                    : "text-gray-600 dark:text-gray-400 hover:text-gray-900 dark:hover:text-white"
-                }`}
-              >
-                All Applications
-              </button>
-            </div>
-            {!myReviewsOnly && (
-              <div className="w-64">
-                <ReviewerFilterDropdown
-                  reviewers={programReviewers}
-                  selectedAddresses={selectedReviewerAddresses}
-                  onChange={setSelectedReviewerAddresses}
-                  isLoading={isLoadingProgramReviewers}
-                  isError={isProgramReviewersError}
-                />
-              </div>
-            )}
-          </div>
-          <div className="flex justify-end space-x-2">
-            <Button
-              onClick={() => {
-                setFilters({});
-                setSearchInput("");
-                setSelectedReviewerAddresses([]);
-                debouncedSearch.cancel(); // Cancel any pending debounced search
-                router.push(pathname, { scroll: false });
-              }}
-              variant="secondary"
-              className="w-fit px-3 py-1 border bg-transparent text-zinc-500 font-medium border-zinc-200 dark:border-zinc-400 dark:text-zinc-400 flex flex-row gap-2"
-            >
-              <FunnelIcon className="w-5 h-5" />
-              Clear Filters
-            </Button>
-
-            <Button
-              onClick={() => handleExport("csv")}
-              variant="secondary"
-              disabled={isExporting}
-              className="w-fit px-3 py-1 border bg-transparent text-zinc-500 font-medium border-zinc-200 dark:border-zinc-400 dark:text-zinc-400 flex flex-row gap-2"
-            >
-              <ArrowDownTrayIcon className="w-5 h-5" />
-              {isExporting ? "Exporting..." : "Export CSV"}
-            </Button>
-          </div>
-        </div>
-      </div>
+      <ApplicationListFilterBar
+        searchInput={searchInput}
+        filters={filters}
+        onSearchChange={handleSearchChange}
+        onFilterChange={handleFilterChange}
+        myReviewsOnly={myReviewsOnly}
+        onMyReviewsOnlyChange={setMyReviewsOnly}
+        programReviewers={programReviewers}
+        selectedReviewerAddresses={selectedReviewerAddresses}
+        onReviewerAddressesChange={setSelectedReviewerAddresses}
+        isLoadingProgramReviewers={isLoadingProgramReviewers}
+        isProgramReviewersError={isProgramReviewersError}
+        onClearFilters={clearFilters}
+        onExport={handleExport}
+        isExporting={isExporting}
+      />
 
       {/* "My Applications": inbox (selectable list + inline detail).
           "All Applications": the original table with infinite scroll. */}
@@ -547,9 +238,7 @@ const ApplicationListWithAPI: FC<IApplicationListWithAPIProps> = ({
           isFetchingNextPage={isFetchingNextPage}
           fetchNextPage={loadMore}
           total={total}
-          hasActiveFilters={Boolean(
-            filters.search || filters.status || filters.dateFrom || filters.dateTo
-          )}
+          hasActiveFilters={hasActiveFilters}
           onApplicationHover={onApplicationHover}
         />
       ) : (
